refactored redux to create more sepration of concerns, quickfix: Login connect error
