+++ src/reducers/authReducers.js
@@ -1,4 +1,4 @@
-export const authReducer = (state = {}, action) => {
+export function reducer (state = {}, action) {
     switch (action.type) {
     case 'LOGIN':
         return {
