+++ src/reducers/index.js
@@ -1,0 +1,6 @@
+import { combineReducers } from 'react-redux';
+import authInfo from './authReducers';
+
+export default combineReducers({
+  authInfo
+})
